requirements.txt
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
docutils==0.16
Flask==1.1.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
lockfile==0.12.2
luigi==3.0.1
MarkupSafe==1.1.1
numpy==1.19.1
pandas==1.1.0
python-daemon==2.2.4
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
tornado==5.1.1
urllib3==1.25.10
Werkzeug==1.0.1
